Intial commit
Initial project structure which includes:

- typescript
- eslint
- jest
- PlaylistNomad class
- test for PlaylistNomad
- basic npm scripts to trigger lint and test

Adds travis for CI

The only thing specified here is that we want to use the nodejs language. Everything else should be done by default. According to travis docs if a yarn.lock file exists, yarn will be used to install, run and test the application instead of npm.